introduce-logprob-result-processor-move: Move logprob assembly to SchedulerLogprobResultProcessor
Mechanical cut + paste for the ``introduce-logprob-result-processor`` mech move.

Cut the logprob @staticmethods (after prep) from
``SchedulerOutputProcessorMixin`` and paste them into
``SchedulerLogprobResultProcessor`` body in
``scheduler_components/logprob_result_processor.py``.

Drop ``@staticmethod`` decorators; simplify
``self: "SchedulerLogprobResultProcessor"`` → bare ``self`` (in class context
the type is implicit). Method bodies otherwise byte-identical.

All callers updated:
  ``self.<m>(self.logprob_result_processor, ...)`` →
  ``self.logprob_result_processor.<m>(...)``
(pure prefix transformation).

Refactor chain ID: introduce-logprob-result-processor-move
